[Ref] Silence warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@16968 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/AbstractVstEditor.cpp
@@ -534,7 +534,7 @@ LRESULT CAbstractVstEditor::OnCustomKeyMsg(WPARAM wParam, LPARAM /*lParam*/)
 		if(ValidateCurrentInstrument())
 		{
 			CModDoc *pModDoc = m_VstPlugin.GetModDoc();
-			const ModCommand::NOTE note = pModDoc->GetNoteWithBaseOctave(wParam - kcVSTGUIStartNotes, m_nInstrument);
+			const ModCommand::NOTE note = pModDoc->GetNoteWithBaseOctave(static_cast<int>(wParam - kcVSTGUIStartNotes), m_nInstrument);
 			if(ModCommand::IsNote(note))
 			{
 				pModDoc->PlayNote(PlayNoteParam(note).Instrument(m_nInstrument), &m_noteChannel);
@@ -547,7 +547,7 @@ LRESULT CAbstractVstEditor::OnCustomKeyMsg(WPARAM wParam, LPARAM /*lParam*/)
 		if(ValidateCurrentInstrument())
 		{
 			CModDoc *pModDoc = m_VstPlugin.GetModDoc();
-			const ModCommand::NOTE note = pModDoc->GetNoteWithBaseOctave(wParam - kcVSTGUIStartNoteStops, m_nInstrument);
+			const ModCommand::NOTE note = pModDoc->GetNoteWithBaseOctave(static_cast<int>(wParam - kcVSTGUIStartNoteStops), m_nInstrument);
 			if(ModCommand::IsNote(note))
 			{
 				pModDoc->NoteOff(note, false, m_nInstrument, m_noteChannel[note - NOTE_MIN]);
